OPT.ENH: move knobspace slice visualization method to base
